telega/services.py
import telegram
from django.conf import settings

bot = telegram.Bot(token=settings.TELEGRAM_TOKEN)


def send_telegram_message(chat_id: str, text: str) -> bool:
    """ Отправляет сообщения пользователю по ID. """

    try:
        bot.send_message(chat_id=chat_id, text=text)
        return True

    except telegram.error.TelegramError as e:
        print(f"Возникла ошибка: {e}")
        return False
