pyproject.toml
[project]
name = "t-e-t"
version = "0.1.0"
description = "Track Everything Today"
authors = [
    { name = "Yurii", email = "yurii.1986.e@gmail.com" }
]

readme = "README.md"

dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "ipython (>=9.3.0,<10.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "python-telegram-bot (>=22.5,<23.0)",
    "coverage (>=7.10.7,<8.0.0)",
]

[tool.poetry.dependencies]
python = "^3.12"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = ".git"

[tool.isort]
line_length = 119
multi_line_output = 3